+++ hp3parclient/client.py
@@ -31,7 +31,7 @@ import time
 
 try:
     # For Python 3.0 and later
-    from urlib.parse import quote
+    from urllib.parse import quote
 except ImportError:
     # Fall back to Python 2's urllib2
     from urllib2 import quote