+++ docs/changelog.rst
@@ -3,8 +3,11 @@ Changelog
 
 Changes in Version 3.1.1
 ------------------------
+
 * Fixed an issue with building the ClientException when body of the response
   was empty.
+* Fixed spelling error in urllib import for Python 3.0 or greater
+  environments.
 
 Changes in Version 3.1.0
 ------------------------